Readme: Use docker volume for data storage
Right now, the metadata permanently resides in the `mysql`, but if you update the `spex` container, the cached PDFs are lost. With a volume, you save that data and the metadata remains intact. I previously had some trouble attaching a fresh `spex` to a mysql DB that contained data from a previous instance.